+++ src/ pages/Main/Banner/Banner.scss
