+++ src/ pages/Main/Collection/Collection.scss
