+++ src/ pages/Main/Main.scss
