+++ src/ pages/Main/Skeleton/Skeleton.js
