+++ src/ pages/Main/Skeleton/Skeleton.scss
