+++ src/ pages/ProductDetail/ProductDetail.scss
