--- src/pages/Login/Login.scss
@@ -1,69 +1,0 @@
-@import '../../styles/variables.scss';
-.formContainer {
-  position: absolute;
-  width: 310px;
-  height: 380px;
-  align-content: center;
-  transform: translate(-50%, -20%);
-  top: 50%;
-  left: 49%;
-
-  .loginTitle {
-    text-align: center;
-  }
-  input {
-    margin: 7px;
-  }
-  h3 {
-    margin: 7px;
-    padding-bottom: 13px;
-  }
-
-  .loginUser {
-    width: 100%;
-    padding: 0px 23px;
-    line-height: 45px;
-    align-content: center;
-    border: 1px solid $light-grey-border;
-    border-radius: 3px;
-    font-size: 14px;
-    &::placeholder {
-      border: 1px $light-grey;
-    }
-  }
-  .buttons {
-    display: block;
-    width: 100%;
-    height: 50px;
-    border-radius: 3px;
-    margin: 9px;
-    color: $main-purple;
-    font-size: 15px;
-    text-align: center;
-    overflow: hidden;
-  }
-  .btnLogin {
-    border: 1px solid $main-purple;
-    background-color: $main-purple;
-    color: $light-grey;
-  }
-  .btnSignUp {
-    border: 1px solid $main-purple;
-    background-color: $light-grey;
-  }
-
-  .loginFind {
-    float: right;
-    padding: 13px 0px 9px 0px;
-    text-align: center;
-    font-size: 13px;
-    .bar {
-      float: inherit;
-      width: 1px;
-      height: 11px;
-      margin: 2px 1px 0 -2px;
-      background-color: $black;
-      position: fixed;
-    }
-  }
-}
