--- src/pages/Signup/Signup.scss
@@ -1,162 +1,0 @@
-@import '../../styles/variables.scss';
-div {
-  font-size: 1rem;
-
-  h2 {
-    line-height: 35px;
-    letter-spacing: 1px;
-    margin-top: 23px;
-    color: $black;
-    font-weight: bold;
-    text-align: center;
-    font-size: 28px;
-  }
-
-  .pageSub {
-    padding-right: 420px;
-    padding-top: 45px;
-    line-height: 11px;
-    color: $dark-grey-text;
-    text-align: right;
-    font-size: 12px;
-    font-weight: bold;
-  }
-
-  .icons {
-    padding-right: 2px;
-    color: #ee6a7b;
-  }
-
-  .tableBar {
-    position: absolute;
-    width: 650px;
-    left: 320px;
-    margin-top: 10px;
-    padding-bottom: 35px;
-    border-top: 1px solid $black;
-    border-bottom: 1px solid $black;
-  }
-  .mapList {
-    display: table-row-group;
-    vertical-align: middle;
-    border-color: inherit;
-  }
-  .mapOne {
-    display: table-row;
-    vertical-align: inherit;
-    border-color: inherit;
-  }
-  .tableMap {
-    width: 159px;
-    padding: 40px 0 0 20px;
-    line-height: 20px;
-    color: $black;
-    font-weight: bold;
-    text-align: left;
-    vertical-align: top;
-  }
-
-  td {
-    margin-left: 40px;
-    padding: 9px;
-    border-top: 0;
-    text-align: left;
-  }
-
-  .inputTests {
-    height: 44px;
-    width: 43%;
-    padding: 0 11px;
-    border: 1px solid $light-grey-border;
-    line-height: 10px;
-    border-radius: 4px;
-    outline: none;
-    vertical-align: top;
-    color: $black;
-    background: $light-grey;
-    font-size: 14px;
-  }
-  .tableBars {
-    position: absolute;
-    width: 790px;
-    left: 490px;
-    margin-top: 26px;
-  }
-
-  .userAgreePart {
-    display: table-row-group;
-    vertical-align: middle;
-    border-color: inherit;
-
-    .agreePart {
-      position: absolute;
-      margin-top: 35px;
-      left: -150px;
-      text-align: left;
-    }
-  }
-
-  input[type='checkbox'] {
-    position: absolute;
-    width: 20px;
-    height: 20px;
-    left: 4px;
-  }
-
-  .inputBackContent {
-    display: flex;
-    flex-direction: column;
-    padding: 24px;
-    height: 33px;
-    margin: 12px;
-    margin-left: -23px;
-    width: 60%;
-    border-radius: 4px;
-    border: 1px solid $light-grey-border;
-    text-align: left;
-  }
-
-  .checkboxAll {
-    margin-left: 29px;
-    margin-top: 33px;
-    font-size: 28px;
-    font-weight: bold;
-    text-align: left;
-  }
-
-  .checkboxAllSubText {
-    font-size: 11px;
-    font-weight: 400;
-  }
-
-  .partMap {
-    display: table;
-    margin-top: 31px;
-    margin-left: 30px;
-    height: 30px;
-    text-align: left;
-    font-size: 18px;
-    font-weight: bold;
-  }
-
-  .partMap1 {
-    margin-top: -150px;
-  }
-  .partMap2 {
-    margin-top: -90px;
-  }
-  .partMap3 {
-    margin-top: -28px;
-  }
-  .memberJoins {
-    position: absolute;
-    border: 1px solid $main-purple;
-    bottom: -90px;
-    left: 130px;
-    width: 240px;
-    height: 56px;
-    color: $light-grey;
-    background-color: $main-purple;
-    font-size: 21px;
-  }
-}
